pyproject.toml
[project]
name = "ultimate-python-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "datetime>=5.5",
    "logging>=0.4.9.6",
    "pathlib>=1.0.1",
    "pydantic[mypy]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "types-pyyaml>=6.0.12.20250516",
    "typing>=3.10.0.0",
    "toml>=0.10.2",
    "types-toml>=0.10.8.20240310",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
    "sourcery>=1.37.0",
]


[tool.pytest.ini_options]
pythonpath="src"

# uv run pytest -s .\tests\test_*.py


[tool.myproject.dirs]
tree = [
    "src/",
    "tests/",
    "outputs/",
    "outputs/{output_folder}/logs/",
    "outputs/{output_folder}/data/",
    "outputs/{output_folder}/configs_used/",
    "config/",
    "config/logging_configs",
    "config.yaml",
    "main.py",
    ".env",
    "pyproject.toml"
]